+++ pint/facets/plain/quantity.py
@@ -11,7 +11,6 @@ from __future__ import annotations
 import bisect
 import copy
 import datetime
-import functools
 import inspect
 import locale
 import math
@@ -42,15 +41,11 @@ from ...compat import (
     HAS_NUMPY,
     _to_magnitude,
     babel_parse,
-    compute,
-    dask_array,
     eq,
     is_duck_array_type,
     is_upcast_type,
     ndarray,
     np,
-    persist,
-    visualize,
     zero_or_nan,
 )
 from ...errors import DimensionalityError, OffsetUnitCalculusError, PintTypeError
@@ -135,20 +130,6 @@ def method_wraps(numpy_func):
         func.__wrapped__ = numpy_func
 
         return func
-
-    return wrapper
-
-
-def check_dask_array(f):
-    @functools.wraps(f)
-    def wrapper(self, *args, **kwargs):
-        if isinstance(self._magnitude, dask_array.Array):
-            return f(self, *args, **kwargs)
-        else:
-            msg = "Method {} only implemented for objects of {}, not {}".format(
-                f.__name__, dask_array.Array, self._magnitude.__class__
-            )
-            raise AttributeError(msg)
 
     return wrapper
 
@@ -1796,93 +1777,6 @@ class PlainQuantity(PrettyIPython, SharedRegistryObject, Generic[_MagnitudeType]
 
     def to_timedelta(self: PlainQuantity[float]) -> datetime.timedelta:
         return datetime.timedelta(microseconds=self.to("microseconds").magnitude)
-
-    # Dask.array.Array ducking
-    def __dask_graph__(self):
-        if isinstance(self._magnitude, dask_array.Array):
-            return self._magnitude.__dask_graph__()
-        else:
-            return None
-
-    def __dask_keys__(self):
-        return self._magnitude.__dask_keys__()
-
-    def __dask_tokenize__(self):
-        from dask.base import tokenize
-
-        return (PlainQuantity, tokenize(self._magnitude), self.units)
-
-    @property
-    def __dask_optimize__(self):
-        return dask_array.Array.__dask_optimize__
-
-    @property
-    def __dask_scheduler__(self):
-        return dask_array.Array.__dask_scheduler__
-
-    def __dask_postcompute__(self):
-        func, args = self._magnitude.__dask_postcompute__()
-        return self._dask_finalize, (func, args, self.units)
-
-    def __dask_postpersist__(self):
-        func, args = self._magnitude.__dask_postpersist__()
-        return self._dask_finalize, (func, args, self.units)
-
-    @staticmethod
-    def _dask_finalize(results, func, args, units):
-        values = func(results, *args)
-        return PlainQuantity(values, units)
-
-    @check_dask_array
-    def compute(self, **kwargs):
-        """Compute the Dask array wrapped by pint.PlainQuantity.
-
-        Parameters
-        ----------
-        **kwargs : dict
-            Any keyword arguments to pass to ``dask.compute``.
-
-        Returns
-        -------
-        pint.PlainQuantity
-            A pint.PlainQuantity wrapped numpy array.
-        """
-        (result,) = compute(self, **kwargs)
-        return result
-
-    @check_dask_array
-    def persist(self, **kwargs):
-        """Persist the Dask Array wrapped by pint.PlainQuantity.
-
-        Parameters
-        ----------
-        **kwargs : dict
-            Any keyword arguments to pass to ``dask.persist``.
-
-        Returns
-        -------
-        pint.PlainQuantity
-            A pint.PlainQuantity wrapped Dask array.
-        """
-        (result,) = persist(self, **kwargs)
-        return result
-
-    @check_dask_array
-    def visualize(self, **kwargs):
-        """Produce a visual representation of the Dask graph.
-
-        The graphviz library is required.
-
-        Parameters
-        ----------
-        **kwargs : dict
-            Any keyword arguments to pass to ``dask.visualize``.
-
-        Returns
-        -------
-
-        """
-        visualize(self, **kwargs)
 
 
 # TODO: Remove in the near future
