+++ pint/registry.py
@@ -37,6 +37,7 @@ from __future__ import annotations
 
 from . import registry_helpers
 from .facets.context import ContextRegistry
+from .facets.dask import DaskRegistry
 from .facets.nonmultiplicative import NonMultiplicativeRegistry
 from .facets.numpy import NumpyRegistry
 from .facets.plain import PlainRegistry
@@ -47,6 +48,7 @@ from .util import logger, pi_theorem
 class UnitRegistry(
     SystemRegistry,
     ContextRegistry,
+    DaskRegistry,
     NumpyRegistry,
     NonMultiplicativeRegistry,
     PlainRegistry,
